compiles with physics enabled again; added new example binaries to .gitignore

+++ include/gua/physics/PhysicsUtils.hpp
@@ -88,10 +88,10 @@ inline btQuaternion const math::extract_quaternion(const math::mat4& m) {
   float tr, s;
   float q[4];
   int i, j, k;
-  float mat[4][4] = { { m.m00, m.m01, m.m02, m.m03 },
-                      { m.m04, m.m05, m.m06, m.m07 },
-                      { m.m08, m.m09, m.m10, m.m11 },
-                      { m.m12, m.m13, m.m14, m.m15 } };
+  float mat[4][4] = { { static_cast<float>(m.m00), static_cast<float>(m.m01), static_cast<float>(m.m02), static_cast<float>(m.m03) },
+                      { static_cast<float>(m.m04), static_cast<float>(m.m05), static_cast<float>(m.m06), static_cast<float>(m.m07) },
+                      { static_cast<float>(m.m08), static_cast<float>(m.m09), static_cast<float>(m.m10), static_cast<float>(m.m11) },
+                      { static_cast<float>(m.m12), static_cast<float>(m.m13), static_cast<float>(m.m14), static_cast<float>(m.m15) } };
 
   int nxt[3] = { 1, 2, 0 };
   tr = mat[0][0] + mat[1][1] + mat[2][2];
